Fix recognition of jquery to also work with non-globals
When jquery is not attached globally, the previous detection for jQuery
would fail and think it would be Zepto, even if jQuery is used.
Since $ is the local variable for jQuery/Zepto, this commit makes the
distinction between jQuery and Zepto work with a non-global jQuery.

$.fn.jquery is undefined in Zepto, but in jQuery it returns the
version number and thus returns true.