Change mfp.updateItemHTML() order
In case we just update the content, we still need to use the new settings in `data` variable. So `mfp.updateItemHTML();` should be called after `mfp.st = $.extend(true, {}, $.magnificPopup.defaults, data );`. (or the ajax request will use the old data)
I am not sure if this change is a good idea but it fixed by problem.