Update core.js
Fix for allow Shift + Click when midClick option is set to false. Shift + Click in browsers opens the href in a new window. Added altKey just as a precaution if a browser somewhere uses it. But if they don't then alt + click will do nothing, so i'm half/half on adding that. But in my production module i did add the altKey thing.