Language direction support for RTL languages
Added a langDir setting. If it's not set specifically, it defaults to that of the document. The left/right buttons and shortcuts now take this into account.

This means that for RTL languages (Hebrew, Arabic) the button on the left is now "next" and the button on the right is now "prev". 

LTR languages aren't affected by these changes.